Make span prediction work
we now predict spans for all span types, not just the first.
our F-score is still suspiciously low -- we hypothesize this has to do with
how we predict, and specifically split, frames.  I made a small fix in an
attempt to address this, but it didn't have the desired effect.  I am
still investigating