Bugfix: Allow calculated values in progress bar
A new, optional "calc" keyword argument to progressify will allow the
special sequence %c to be replaced by the eval()'ed version of a string.
This allows for dynamic looking progress bar labels. Note that only
variables outside of the loop can be directly handed in, except for
special values obtained by replacements, such as %i and %e.

The list of special sequences is easily extensible now.