Replace calc string with a callback
eval-ing the provided calc string seemed a bit messy, and it seemed it would be
a lot of mental load to write a string that would have %-patterns replaced and then be evaluated.

Instead, I allowed the string (now called message) patter to either be a string, in which case the
old behavior is maintained, or a callable returning a string, which will be called with the given
index and element.  This allows callers to provide arbitrary logic, but in the form of a callable, instead
of a string to be eval-ed, which seems to be a better abstraction imo